eliminate coverity warning

+++ coders/tiff.c
@@ -1897,7 +1897,7 @@ static Image *ReadTIFFImage(const ImageInfo *image_info,
           Convert stripped TIFF image.
         */
         extent=4*MagickMax(image->columns*(samples_per_pixel+extra_samples)*
-          (image->depth+7)/8,TIFFStripSize(tiff));
+          (image->depth+7)/8,(size_t) TIFFStripSize(tiff));
         strip_pixels=(unsigned char *) AcquireQuantumMemory(extent,
           sizeof(*strip_pixels));
         if (strip_pixels == (unsigned char *) NULL)
